etl/src/requirements.txt
pandas==1.4.3
google-cloud-BigQuery==3.3.0
requests==2.28.1
pytest==7.1.2
mock==4.0.3